tests/wpt/mozilla/tests/mozilla/track_line.html

<!doctype html>
<meta charset="utf-8">
<title>Test that errors from inline scripts report meaningful line numbers</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({allow_uncaught_exception:true});
var t = async_test("error event has proper line number");
var errors = 0;
var expected_lines = [[location.href, 27],
                      [location.href, 31],
                      [location.href.replace('track_line.html', 'resources/external.js'), 4]];
window.addEventListener('error', t.step_func(function(e) {
    assert_true(e instanceof ErrorEvent);
    var expected = expected_lines[errors];
    assert_equals(e.filename, expected[0]);
    assert_equals(e.lineno, expected[1]);
    errors++;
    if (errors == expected_lines.length) {
        t.done();
    }
}), true);
</script>
<div id="testdiv"></div>
<script>
  var div = document.querySelector('#testdiv');
  div.setAttribute('onclick', 'this_is_a_js_error');
  div.dispatchEvent(new Event('click'));
</script>
<script>
this_is_a_js_error
</script>
<script src="resources/external.js"></script>
